# target/classes/srs-openapi.yaml
openapi: "3.0.3"
info:
  title: "srs API"
  description: "srs API"
  version: "1.0.0"
servers:
  - url: "https://srs"
paths:
...
  /cells/add:
    post:
      summary: "POST cells/add"
      operationId: "addCell"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /cells/delete/{id}:
    delete:
      summary: "DELETE cells/delete/{id}"
      operationId: "deleteCell"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /cells/filter:
    get:
      summary: "GET cells/filter"
      operationId: "getCellsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CellDTO"
  /cells/update:
    put:
      summary: "PUT cells/update"
      operationId: "updateCell"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /departments/add:
    post:
      summary: "POST departments/add"
      operationId: "addDepartment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /departments/delete/{id}:
    delete:
      summary: "DELETE departments/delete/{id}"
      operationId: "deleteDepartment"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /departments/filter:
    get:
      summary: "GET departments/filter"
      operationId: "getDepartmentsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DepartmentDTO"
  /departments/update:
    put:
      summary: "PUT departments/update"
      operationId: "updateDepartment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /permissions/add:
    post:
      summary: "POST permissions/add"
      operationId: "addPermission"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /permissions/delete/{id}:
    delete:
      summary: "DELETE permissions/delete/{id}"
      operationId: "deletePermission"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /permissions/filter:
    get:
      summary: "GET permissions/filter"
      operationId: "getPermissionsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PermissionDTO"
  /permissions/update:
    put:
      summary: "PUT permissions/update"
      operationId: "updatePermission"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /priorities/add:
    post:
      summary: "POST priorities/add"
      operationId: "addPriority"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriorityDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /priorities/delete/{id}:
    delete:
      summary: "DELETE priorities/delete/{id}"
      operationId: "deletePriority"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /priorities/filter:
    get:
      summary: "GET priorities/filter"
      operationId: "getPrioritiesByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriorityFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PriorityDTO"
  /priorities/update:
    put:
      summary: "PUT priorities/update"
      operationId: "updatePriority"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriorityDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
components:
  schemas:
    CellDTO:
      type: "object"
      properties:
        cellId:
          type: "integer"
          format: "int32"
        prisonId:
          type: "integer"
          format: "int32"
        segment:
          type: "string"
        blok:
          type: "string"
        assignDate:
          type: "string"
          format: "date"
        unassignDate:
          type: "string"
          format: "date"
...
    OperationStatus:
      type: "object"
      properties:
        targetClassName:
          type: "string"
        recordId:
          type: "integer"
          format: "int32"
        operationName:
          type: "string"
        success:
          type: "boolean"
        msgList:
          type: "array"
          items:
            type: "string"
    CellFilter:
      type: "object"
      properties: { }
    DepartmentDTO:
      type: "object"
      properties:
        departmentId:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        description:
          type: "string"
    DepartmentFilter:
      type: "object"
      properties: { }
    PermissionDTO:
      type: "object"
      properties:
        permissionId:
          type: "integer"
          format: "int32"
        categoryId:
          type: "integer"
          format: "int32"
        allowRead:
          type: "boolean"
        allowAdd:
          type: "boolean"
        allowModify:
          type: "boolean"
        allowDelete:
          type: "boolean"
        classId:
          type: "integer"
          format: "int32"
    PermissionFilter:
      type: "object"
      properties: { }
    PriorityDTO:
      type: "object"
      properties:
        priorityId:
          type: "integer"
          format: "int32"
        type:
          type: "string"
        description:
          type: "string"
    PriorityFilter:
      type: "object"
      properties: { }
...